requirements.txt
django~=3.2.4
djangorestframework~=3.12.4
psycopg2~=2.8.6
coreapi~=2.3.3
django-import-export~=2.5.0